pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fuel-prediction"
version = "1.0.0"
description = "Aircraft fuel burn prediction using hybrid stacking (GBM + LSTM)"
readme = "README.md"
authors = [{name = "Team Sincere Glacier"}]
license = {text = "GNU GPLv3"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.8"
dependencies = [
    "torch>=2.0.0",
    "lightgbm>=4.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "joblib>=1.3.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "click>=8.1.0",
    "matplotlib>=3.7.0",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.5.0",
]

[project.scripts]
fuel-predict = "fuel_prediction.cli:cli"

[project.urls]
Homepage = "https://github.com/johntad110/sincere-glacier-prc2025"
Repository = "https://github.com/johntad110/sincere-glacier-prc2025"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
fuel_prediction = ["config/*.yaml"]
